Refactor forcefield. Now theres a Base
