--- Runtime/Scripts/Force/IForceField.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace Force
-{
-    public interface IForceField
-    {
-        public Vector3 GetForceAt(Vector3 position);
-    }
-}
